backend/requirements.txt
fastapi==0.115.12
uvicorn>=0.31.1
sqlalchemy>=2.0.36
psycopg2-binary==2.9.11
alembic==1.13.1
python-dotenv==1.0.1
pydantic==2.11.5
pydantic-settings==2.9.1
httpx>=0.27.1
openai>=2.8.0
cryptography==46.0.3
agent-framework-github-copilot>=1.0.0b260319
python-multipart==0.0.22
python-docx==1.2.0
pypdf==4.3.1
duckduckgo-search==6.3.7
tavily-python>=0.5.0
beautifulsoup4==4.13.5
celery==5.4.0
redis==5.2.1
celery[redis]==5.4.0
croniter==2.0.1
watchdog>=4.0.0
